flux metering: the ThorCam settings are integers, and say why no camera was found
Two fixes, both found by finally running against real hardware.

`flux_gain` was typed `float`. The SDK's setters are `c_int` throughout --
`tl_camera_set_gain` and `tl_camera_set_black_level` both take `c_int`, and the
getters return `int(...)` -- so a float reaches ctypes and raises
`TypeError: int expected` rather than being rounded. The default of 0 worked by
luck, being already an integer; `flux_gain=0.5` would have failed inside a run
that had already claimed the mount. The CS165MU's ranges, read from the camera:
gain (0, 480), black level (0, 511), exposure (64, 26843418) us.

"no Thorlabs camera was found" was too vague to act on, and cost a real
diagnosis. The SDK reports an empty list identically for "unplugged", "no driver"
and "another process has it". The first time this fired, the cause was the
second: the camera enumerated as a bare `TSI` device with problem code 28,
CM_PROB_FAILED_INSTALL, because the Python wheel ships the user-mode DLLs but not
the kernel driver. The message now names all three possibilities and tells the
reader where to look.

For the record, the working case: `Thorlabs Camera Zelux`, status OK, service
CYUSB3, driver 1.2.3.14 -- and a full ThorCam cycle (open, configure, two
exposures, close) then ran against the real camera on its first attempt.

+++ src/flux_metering/session.py
@@ -105,7 +105,7 @@ class FluxMeteringParams:
     max_rings: int = 6
     patience_rings: int = 1
     max_radius_arcsec: float = 10.0
-    flux_gain: float = 0.0
+    flux_gain: int = 0
     flux_black_level: int = 3
     number_of_frames: int = 3
     usable_fraction: float = 0.66
